Migration implementation and deployment improvements

+++ src/AgentKeyWhitelist.sol
@@ -2,15 +2,7 @@
 pragma solidity ^0.8.13;
 
 contract AgentKeyWhitelist {
-    function authorizeTransfer(
-        address _from,
-        address _to,
-        uint,
-        bool
-    ) external pure {
-        require(
-            _from == address(0) || _to == address(0),
-            "TRANSFERS_DISABLED"
-        );
+    function authorizeTransfer(address _from, address _to, uint256, bool) external pure {
+        require(_from == address(0) || _to == address(0), "TRANSFERS_DISABLED");
     }
 }
